backend/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
